cleanup for att protection
