+++ validator/db/kv/attester_protection.go
@@ -9,6 +9,7 @@ import (
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"github.com/prysmaticlabs/prysm/shared/slashutil"
 	bolt "go.etcd.io/bbolt"
+	"go.opencensus.io/trace"
 )
 
 // SlashingKind used for helpful information upon detection.
@@ -121,6 +122,107 @@ func (store *Store) ApplyAttestationForPubKey(
 		if err != nil {
 			return err
 		}
-		return sourceEpochsBucket.Put(sourceEpochBytes, targetEpochBytes)
+		if err := sourceEpochsBucket.Put(sourceEpochBytes, targetEpochBytes); err != nil {
+			return err
+		}
+
+		// Initialize buckets for the lowest target and source epochs.
+		lowestSourceBucket, err := tx.CreateBucketIfNotExists(lowestSignedSourceBucket)
+		if err != nil {
+			return err
+		}
+		lowestTargetBucket, err := tx.CreateBucketIfNotExists(lowestSignedTargetBucket)
+		if err != nil {
+			return err
+		}
+
+		// If the incoming source epoch is lower than the lowest signed source epoch, override.
+		lowestSignedSourceBytes := lowestSourceBucket.Get(pubKey[:])
+		var lowestSignedSourceEpoch uint64
+		if len(lowestSignedSourceBytes) >= 8 {
+			lowestSignedSourceEpoch = bytesutil.BytesToUint64BigEndian(lowestSignedSourceBytes)
+		}
+		if len(lowestSignedSourceBytes) == 0 || att.Data.Source.Epoch < lowestSignedSourceEpoch {
+			if err := lowestSourceBucket.Put(
+				pubKey[:], bytesutil.Uint64ToBytesBigEndian(att.Data.Source.Epoch),
+			); err != nil {
+				return err
+			}
+		}
+
+		// If the incoming target epoch is lower than the lowest signed target epoch, override.
+		lowestSignedTargetBytes := lowestTargetBucket.Get(pubKey[:])
+		var lowestSignedTargetEpoch uint64
+		if len(lowestSignedTargetBytes) >= 8 {
+			lowestSignedTargetEpoch = bytesutil.BytesToUint64BigEndian(lowestSignedTargetBytes)
+		}
+		if len(lowestSignedTargetBytes) == 0 || att.Data.Target.Epoch < lowestSignedTargetEpoch {
+			if err := lowestTargetBucket.Put(
+				pubKey[:], bytesutil.Uint64ToBytesBigEndian(att.Data.Target.Epoch),
+			); err != nil {
+				return err
+			}
+		}
+		return nil
 	})
+}
+
+// AttestedPublicKeys retrieves all public keys in our attestation history bucket.
+func (store *Store) AttestedPublicKeys(ctx context.Context) ([][48]byte, error) {
+	ctx, span := trace.StartSpan(ctx, "Validator.AttestedPublicKeys")
+	defer span.End()
+	var err error
+	attestedPublicKeys := make([][48]byte, 0)
+	err = store.view(func(tx *bolt.Tx) error {
+		bucket := tx.Bucket(historicAttestationsBucket)
+		return bucket.ForEach(func(key []byte, _ []byte) error {
+			pubKeyBytes := [48]byte{}
+			copy(pubKeyBytes[:], key)
+			attestedPublicKeys = append(attestedPublicKeys, pubKeyBytes)
+			return nil
+		})
+	})
+	return attestedPublicKeys, err
+}
+
+// LowestSignedSourceEpoch returns the lowest signed source epoch for a validator public key.
+// If no data exists, returning 0 is a sensible default.
+func (store *Store) LowestSignedSourceEpoch(ctx context.Context, publicKey [48]byte) (uint64, error) {
+	ctx, span := trace.StartSpan(ctx, "Validator.LowestSignedSourceEpoch")
+	defer span.End()
+
+	var err error
+	var lowestSignedSourceEpoch uint64
+	err = store.view(func(tx *bolt.Tx) error {
+		bucket := tx.Bucket(lowestSignedSourceBucket)
+		lowestSignedSourceBytes := bucket.Get(publicKey[:])
+		// 8 because bytesutil.BytesToUint64BigEndian will return 0 if input is less than 8 bytes.
+		if len(lowestSignedSourceBytes) < 8 {
+			return nil
+		}
+		lowestSignedSourceEpoch = bytesutil.BytesToUint64BigEndian(lowestSignedSourceBytes)
+		return nil
+	})
+	return lowestSignedSourceEpoch, err
+}
+
+// LowestSignedTargetEpoch returns the lowest signed target epoch for a validator public key.
+// If no data exists, returning 0 is a sensible default.
+func (store *Store) LowestSignedTargetEpoch(ctx context.Context, publicKey [48]byte) (uint64, error) {
+	ctx, span := trace.StartSpan(ctx, "Validator.LowestSignedTargetEpoch")
+	defer span.End()
+
+	var err error
+	var lowestSignedTargetEpoch uint64
+	err = store.view(func(tx *bolt.Tx) error {
+		bucket := tx.Bucket(lowestSignedTargetBucket)
+		lowestSignedTargetBytes := bucket.Get(publicKey[:])
+		// 8 because bytesutil.BytesToUint64BigEndian will return 0 if input is less than 8 bytes.
+		if len(lowestSignedTargetBytes) < 8 {
+			return nil
+		}
+		lowestSignedTargetEpoch = bytesutil.BytesToUint64BigEndian(lowestSignedTargetBytes)
+		return nil
+	})
+	return lowestSignedTargetEpoch, err
 }
